requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.31
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
smmap==5.0.0
Werkzeug==2.3.6
zipp==3.15.0
